requirements.txt
aiogram
python-dotenv
transformers